add static_cast to fix code scanning warning

+++ test/sampling_tests.cpp
@@ -262,7 +262,7 @@ void RunSamplingTest(int batch_size, int k, float p, int vocab_size, int num_ite
 
   std::vector<float> logits_cpu(static_cast<size_t>(vocab_size) * batch_size);
   for (int b = 0; b < batch_size; b++) {
-    std::copy(single_logits.begin(), single_logits.end(), logits_cpu.begin() + b * vocab_size);
+    std::copy(single_logits.begin(), single_logits.end(), logits_cpu.begin() + static_cast<ptrdiff_t>(b) * vocab_size);
   }
 
   // --- 3. Pre-compute the Expected Distribution ---
